requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dropbox==12.0.2
Flask==3.1.0
flask-swagger-ui==4.11.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector==2.2.9
packaging==24.2
ply==3.11
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
stone==3.3.1
urllib3==2.3.0
Werkzeug==3.1.3
